better layout support
- use metadata to set layout
- cleanup rehype-estree
