+++ packages/rehype-estree/src/index.ts
@@ -10,6 +10,16 @@ function rehype2estree(options: Partial<Options> = {}): Transformer {
 
   function transformer(node, file) {
     const _options = { ...DEFAULT_OPTIONS, ...options }
+    const layout = node.data.layout
+    if (typeof layout === 'string') {
+      const _layout = layout.trim().replace(/["']/g, '')
+      if (['', 'nil', 'undefined', 'null'].includes(_layout)) {
+        _options.defaultLayout = undefined
+      } else {
+        _options.defaultLayout = layout
+      }
+    }
+
     const estree = toEstree(node, _options)
     return processEstree(estree, _options)
   }
