+++ packages/rehype-estree/src/process-estree.ts
@@ -1,41 +1,28 @@
-import { SimpleLiteral } from 'estree'
 import { walk } from 'estree-walker'
 import { analyze } from 'periscopic'
 import { Options } from './options'
 import {
-  isDeclaration,
-  isExportDefaultDeclaration,
-  isExportNamedDeclaration,
-  isJSXElement,
-  isJSXExpression,
-} from './type-check'
+  builders as b,
+  checkers as check,
+  createContent,
+  createMakeShortcodeHelper,
+  layoutBuilder,
+} from './trees'
 
 function processEstree(estree, options: Options) {
   const { skipExport, wrapExport, defaultLayout } = options
 
-  let layout: any
   let children = []
-  let mdxLayoutDefault: any
+  const lb = layoutBuilder()
 
   if (defaultLayout) {
-    mdxLayoutDefault = {
-      local: { type: 'Identifier', name: 'default' },
-      source: {
-        type: 'Literal',
-        value: defaultLayout,
-        raw: `'${defaultLayout}'`,
-      },
-    }
+    lb.setDefaultLayout(defaultLayout)
   }
 
   // Find the `export default`, the JSX expression, and leave the rest
   // (import/exports) as they are.
   estree.body = estree.body.filter((child) => {
-    if (
-      child.type === 'ExpressionStatement' &&
-      (child.expression.type === 'JSXFragment' ||
-        child.expression.type === 'JSXElement')
-    ) {
+    if (check.isExpressionStatement(child)) {
       children =
         child.expression.type === 'JSXFragment'
           ? child.expression.children
@@ -55,11 +42,7 @@ function processEstree(estree, options: Options) {
     ...analyze(estree).scope.declarations.keys(),
   ]
 
-  estree.body = [
-    ...estree.body,
-    // ...createMdxLayout(layout, mdxLayoutDefault),
-    ...createMdxContent(children),
-  ]
+  estree.body = [...estree.body, ...createContent(children)]
 
   const components = []
   // Add `orgaType`, `parentName` props to JSX elements.
@@ -72,8 +55,8 @@ function processEstree(estree, options: Options) {
       // ```js
       // export default a = 1, this is the layout
       // ```
-      if (isExportDefaultDeclaration(node)) {
-        layout = node.declaration
+      if (check.isExportDefaultDeclaration(node)) {
+        lb.declareLayout(node.declaration)
         this.remove()
         return
       }
@@ -85,10 +68,10 @@ function processEstree(estree, options: Options) {
       // export {a as default} from "b"
       // export {a as default, b} from "c"
       // ```
-      if (isExportNamedDeclaration(node) && !!node.source) {
+      if (check.isExportNamedDeclaration(node) && !!node.source) {
         node.specifiers = node.specifiers.filter((specifier) => {
           if (specifier.exported.name === 'default') {
-            mdxLayoutDefault = { local: specifier.local, source: node.source }
+            lb.setDefaultLayout({ local: specifier.local, source: node.source })
             return false
           }
 
@@ -103,13 +86,13 @@ function processEstree(estree, options: Options) {
         }
       }
 
-      if (isExportNamedDeclaration(node)) {
+      if (check.isExportNamedDeclaration(node)) {
         declarations.push(node)
         this.remove()
         return
       }
 
-      if (isDeclaration(node)) {
+      if (check.isDeclaration(node)) {
         const names = analyze(node).scope.declarations.keys()
         const clean = [...names].filter((n) => n !== 'MDXContent')
         if (clean.length > 0) {
@@ -120,56 +103,21 @@ function processEstree(estree, options: Options) {
         }
       }
 
-      // if (isExpressionStatement(node)) {
-      //   const exp = node.expression
-      //   let removeNode = false
-      //   if (isJSXFragment(exp)) {
-      //     // @ts-ignore
-      //     children = exp.children
-      //     removeNode = true
-      //   }
-      //   if (isJSXElement(exp)) {
-      //     children = [exp]
-      //     removeNode = true
-      //   }
-
-      //   if (removeNode) {
-      //     this.remove()
-      //     return
-      //   }
-      // }
-
-      if (isJSXElement(node)) {
+      if (check.isJSXElement(node)) {
         const name = node.openingElement.name.name
 
         if (stack.length > 1) {
           const parentName = stack[stack.length - 1]
-
-          node.openingElement.attributes.push({
-            type: 'JSXAttribute',
-            name: { type: 'JSXIdentifier', name: 'parentName' },
-            value: {
-              type: 'Literal',
-              value: parentName,
-              raw: JSON.stringify(parentName),
-            } as SimpleLiteral,
-          })
+          node.openingElement.attributes.push(
+            b.jsxAttribute('parentName', parentName)
+          )
         }
 
         const head = name.charAt(0)
 
         // A component.
         if (head === head.toUpperCase() && name !== 'MDXLayout') {
-          node.openingElement.attributes.push({
-            type: 'JSXAttribute',
-            name: { type: 'JSXIdentifier', name: 'orgaType' },
-            value: {
-              type: 'Literal',
-              value: name,
-              raw: JSON.stringify(name),
-            } as SimpleLiteral,
-          })
-
+          node.openingElement.attributes.push(b.jsxAttribute('orgaType', name))
           if (!inTopScope.includes(name) && !magicShortcodes.includes(name)) {
             magicShortcodes.push(name)
           }
@@ -180,11 +128,11 @@ function processEstree(estree, options: Options) {
     },
     leave: function (node) {
       // unwrap JSXElement
-      if (isJSXExpression(node)) {
+      if (check.isExpressionStatement(node)) {
         this.replace(node.expression)
       }
 
-      if (isJSXElement(node)) {
+      if (check.isJSXElement(node)) {
         stack.pop()
       }
     },
@@ -212,240 +160,13 @@ function processEstree(estree, options: Options) {
   estree.body = [
     ...createMakeShortcodeHelper(magicShortcodes, true),
     ...declarations,
-    ...createMdxLayout(layout, mdxLayoutDefault),
+    ...lb.getLayout(),
     ...components,
     ...estree.body,
     ...exports,
   ]
 
   return estree
-}
-
-function createMdxContent(children) {
-  return [
-    {
-      type: 'FunctionDeclaration',
-      id: { type: 'Identifier', name: 'MDXContent' },
-      expression: false,
-      generator: false,
-      async: false,
-      params: [
-        {
-          type: 'ObjectPattern',
-          properties: [
-            {
-              type: 'Property',
-              method: false,
-              shorthand: true,
-              computed: false,
-              key: { type: 'Identifier', name: 'components' },
-              kind: 'init',
-              value: { type: 'Identifier', name: 'components' },
-            },
-            {
-              type: 'RestElement',
-              argument: { type: 'Identifier', name: 'props' },
-            },
-          ],
-        },
-      ],
-      body: {
-        type: 'BlockStatement',
-        body: [
-          {
-            type: 'ReturnStatement',
-            argument: {
-              type: 'JSXElement',
-              openingElement: {
-                type: 'JSXOpeningElement',
-                attributes: [
-                  {
-                    type: 'JSXAttribute',
-                    name: { type: 'JSXIdentifier', name: 'components' },
-                    value: {
-                      type: 'JSXExpressionContainer',
-                      expression: { type: 'Identifier', name: 'components' },
-                    },
-                  },
-                  {
-                    type: 'JSXSpreadAttribute',
-                    argument: { type: 'Identifier', name: 'props' },
-                  },
-                ],
-                name: { type: 'JSXIdentifier', name: 'MDXLayout' },
-                selfClosing: false,
-              },
-              closingElement: {
-                type: 'JSXClosingElement',
-                name: { type: 'JSXIdentifier', name: 'MDXLayout' },
-              },
-              children: children,
-            },
-          },
-        ],
-      },
-    },
-    {
-      type: 'ExpressionStatement',
-      expression: {
-        type: 'AssignmentExpression',
-        operator: '=',
-        left: {
-          type: 'MemberExpression',
-          object: { type: 'Identifier', name: 'MDXContent' },
-          property: { type: 'Identifier', name: 'isMDXComponent' },
-          computed: false,
-          optional: false,
-        },
-        right: { type: 'Literal', value: true, raw: 'true' },
-      },
-    },
-  ]
-}
-
-function createMdxLayout(declaration, mdxLayoutDefault) {
-  const id = { type: 'Identifier', name: 'MDXLayout' }
-  const init = { type: 'Literal', value: 'wrapper', raw: '"wrapper"' }
-
-  return [
-    mdxLayoutDefault
-      ? {
-          type: 'ImportDeclaration',
-          specifiers: [
-            mdxLayoutDefault.local.name === 'default'
-              ? { type: 'ImportDefaultSpecifier', local: id }
-              : {
-                  type: 'ImportSpecifier',
-                  imported: mdxLayoutDefault.local,
-                  local: id,
-                },
-          ],
-          source: {
-            type: 'Literal',
-            value: mdxLayoutDefault.source.value,
-            raw: mdxLayoutDefault.source.raw,
-          },
-        }
-      : {
-          type: 'VariableDeclaration',
-          declarations: [
-            { type: 'VariableDeclarator', id: id, init: declaration || init },
-          ],
-          kind: 'const',
-        },
-  ]
-}
-
-function createMakeShortcodeHelper(names, useElement) {
-  const func = {
-    type: 'VariableDeclaration',
-    declarations: [
-      {
-        type: 'VariableDeclarator',
-        id: { type: 'Identifier', name: 'makeShortcode' },
-        init: {
-          type: 'ArrowFunctionExpression',
-          id: null,
-          expression: true,
-          generator: false,
-          async: false,
-          params: [{ type: 'Identifier', name: 'name' }],
-          body: {
-            type: 'ArrowFunctionExpression',
-            id: null,
-            expression: false,
-            generator: false,
-            async: false,
-            params: [{ type: 'Identifier', name: 'props' }],
-            body: {
-              type: 'BlockStatement',
-              body: [
-                {
-                  type: 'ExpressionStatement',
-                  expression: {
-                    type: 'CallExpression',
-                    callee: {
-                      type: 'MemberExpression',
-                      object: { type: 'Identifier', name: 'console' },
-                      property: { type: 'Identifier', name: 'warn' },
-                      computed: false,
-                      optional: false,
-                    },
-                    arguments: [
-                      {
-                        type: 'Literal',
-                        value:
-                          'Component `%s` was not imported, exported, or provided by MDXProvider as global scope',
-                      },
-                      { type: 'Identifier', name: 'name' },
-                    ],
-                  },
-                },
-                {
-                  type: 'ReturnStatement',
-                  argument: useElement
-                    ? {
-                        type: 'JSXElement',
-                        openingElement: {
-                          type: 'JSXOpeningElement',
-                          attributes: [
-                            {
-                              type: 'JSXSpreadAttribute',
-                              argument: { type: 'Identifier', name: 'props' },
-                            },
-                          ],
-                          name: { type: 'JSXIdentifier', name: 'div' },
-                          selfClosing: true,
-                        },
-                        closingElement: null,
-                        children: [],
-                      }
-                    : {
-                        type: 'JSXFragment',
-                        openingFragment: { type: 'JSXOpeningFragment' },
-                        closingFragment: { type: 'JSXClosingFragment' },
-                        children: [
-                          {
-                            type: 'JSXExpressionContainer',
-                            expression: {
-                              type: 'MemberExpression',
-                              object: { type: 'Identifier', name: 'props' },
-                              property: {
-                                type: 'Identifier',
-                                name: 'children',
-                              },
-                              computed: false,
-                            },
-                          },
-                        ],
-                      },
-                },
-              ],
-            },
-          },
-        },
-      },
-    ],
-    kind: 'const',
-  }
-
-  const shortcodes = names.map((name) => ({
-    type: 'VariableDeclaration',
-    declarations: [
-      {
-        type: 'VariableDeclarator',
-        id: { type: 'Identifier', name: String(name) },
-        init: {
-          type: 'CallExpression',
-          callee: { type: 'Identifier', name: 'makeShortcode' },
-          arguments: [{ type: 'Literal', value: String(name) }],
-        },
-      },
-    ],
-    kind: 'const',
-  }))
-
-  return shortcodes.length > 0 ? [func, ...shortcodes] : []
 }
 
 export default processEstree
